+++ analysis/import/SourceCodeParser/src/main/kotlin/de/maibornwolff/codecharta/importer/sourcecodeparser/core/domain/metrics/DetailedMetricType.kt
@@ -1,0 +1,5 @@
+package de.maibornwolff.codecharta.importer.sourcecodeparser.core.domain.metrics
+
+enum class DetailedMetricType {
+    LoC, RLoc, MCC
+}