+++ analysis/import/SourceCodeParser/src/main/kotlin/de/maibornwolff/codecharta/importer/sourcecodeparser/oop/domain/tagging/NonCodeTags.kt
@@ -1,0 +1,7 @@
+package de.maibornwolff.codecharta.importer.sourcecodeparser.oop.domain.tagging
+
+import de.maibornwolff.codecharta.importer.sourcecodeparser.core.domain.tagging.Tags
+
+enum class NonCodeTags : Tags {
+    COMMENT, WHITESPACE
+}