+++ analysis/import/SourceCodeParser/src/main/kotlin/de/maibornwolff/codecharta/importer/sourcecodeparser/orchestration/domain/metrics/DefaultOverviewMetricCalculationStrategy.kt
@@ -1,0 +1,14 @@
+package de.maibornwolff.codecharta.importer.sourcecodeparser.orchestration.domain.metrics
+
+import de.maibornwolff.codecharta.importer.sourcecodeparser.core.domain.metrics.DetailedMetricMap
+import de.maibornwolff.codecharta.importer.sourcecodeparser.core.domain.metrics.MetricCalculationStrategy
+import de.maibornwolff.codecharta.importer.sourcecodeparser.core.domain.metrics.OverviewMetricCalculationStrategy
+import de.maibornwolff.codecharta.importer.sourcecodeparser.core.domain.metrics.OverviewMetricMap
+import de.maibornwolff.codecharta.importer.sourcecodeparser.core.domain.tagging.Line
+
+class DefaultOverviewMetricCalculationStrategy : OverviewMetricCalculationStrategy {
+    override fun toOverviewMetrics(detailedMetricMap: DetailedMetricMap?): OverviewMetricMap {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+
+}