+++ analysis/import/SourceCodeParser/src/test/kotlin/de/maibornwolff/codecharta/importer/sourcecodeparser/test_helpers/StreamHelper.kt
@@ -1,0 +1,13 @@
+package de.maibornwolff.codecharta.importer.sourcecodeparser.test_helpers
+
+import java.io.ByteArrayOutputStream
+import java.io.PrintStream
+
+fun retrieveStreamAsString(aMethod: (printStream: PrintStream) -> Unit): String {
+    val baOutputStream = ByteArrayOutputStream()
+    val printStream = PrintStream(baOutputStream)
+
+    aMethod(printStream)
+
+    return baOutputStream.toString()
+}