+++ analysis/import/SourceCodeParser/src/test/kotlin/de/maibornwolff/codecharta/importer/sourcecodeparser/test_helpers/TabularOutputHelper.kt
@@ -1,0 +1,3 @@
+package de.maibornwolff.codecharta.importer.sourcecodeparser.test_helpers
+
+fun elementsOf(text: String) = text.split(' ').filter { it.isNotEmpty() }