+++ visualization/app/codeCharta/core/scenario/scenario.spec.ts
@@ -32,7 +32,6 @@ describe("app.codeCharta.core.scenarioService", function () {
                 "deltaColorFlipped": false,
                 "deltas": false,
                 "heightMetric": "Average Complexity*",
-                "invertHeight": false,
                 "margin": 1,
                 "maximizeDetailPanel": false,
                 "neutralColorRange":
@@ -41,14 +40,18 @@ describe("app.codeCharta.core.scenarioService", function () {
                         "to": 40
                     },
                 "scaling":
-                    {"x": 1,
+                    {
+                        "x": 1,
                         "y": 1,
                         "z": 1
                     },
                 "showDependencies": true,
+                "invertHeight": false,
+                "dynamicMargin": true,
+                "isWhiteBackground": false,
                 "useCouplingHeight": true,
                 "intelligentTemporalCouplingFilter": true,
-                "minimumAverageRevs": 15
+                "minimumAverageRevs": 15,
             },
 
         "autoFitCamera": true
@@ -70,6 +73,13 @@ describe("app.codeCharta.core.scenarioService", function () {
     it("should apply the settings from a given scenario", () => {
         scenarioService.applyScenario(scenario);
         expect(settingsService.settings).toBe(scenario.settings);
+    });
+
+    it("should apply the settings from a given scenario once when applyScenarioOnce is called", () => {
+        scenarioService.applyScenario = jest.fn();
+        scenarioService.applyScenarioOnce(scenario);
+        scenarioService.applyScenarioOnce(scenario);
+        expect(scenarioService.applyScenario).toHaveBeenCalledTimes(1);
     });
 
     it("default scenario should be rloc/mcc/mcc", () => {
@@ -113,6 +123,10 @@ describe("app.codeCharta.core.scenarioService", function () {
                 }
             };
             expect(scenarioService.isScenarioPossible(scenario, metrics)).toBe(false);
+        });
+
+        it("should be impossible when params are null", () => {
+            expect(scenarioService.isScenarioPossible(null, null)).toBe(false);
         });
 
     });